fixed conflicts

+++ Chess_utility.h
@@ -49,12 +49,12 @@ struct Move_list {
 		//delete(move_list);
 	}
 	void clear() {last = move_list;}
-	void add_move(const uint16_t move) { *last++ = move; }
-	uint16_t to_sq(const uint16_t move) const { return move & 0x3F; }
-	uint16_t from_sq(const uint16_t move) const { return (move >> 6) & 0x3F; }
-	uint16_t flags(const uint16_t move) const { return move >> 12; }
-	uint16_t move_list[60];
-	uint16_t* last = move_list; //60 should be enough space for all moves
+	void add_move(const uint32_t move) { *last++ = move; }
+	uint16_t to_sq(const uint32_t move) const { return move & 0x3F; }
+	uint16_t from_sq(const uint32_t move) const { return (move >> 6) & 0x3F; }
+	uint16_t flags(const uint32_t move) const { return move >> 12; }
+	uint32_t move_list[60];
+	uint32_t* last = move_list; //60 should be enough space for all moves
 };
 
 struct Position_list {
